api/requirements.txt
Django
psycopg2-binary
gunicorn
django-cors-headers
django-rest-framework
django-rest-auth
django-allauth
django-registration
djangorestframework_simplejwt
PyJWT
python-dateutil
whitenoise
pytest-django
selenium
django-extensions
# celery==4.4.7
# redis
# django-celery-beat
# django-celery-results
boto3
